Add #09
Remove unfinished one
FInish old algorithm
Add new efficient algorith
